Starting work on better gating modeling
In adding additional tests for CheckCount, I realized how insufficient the data/modeling was for correctly and generically getting accurate counts of checks to find an item.

This changes the GatedAttribute to have up to two gating items which will be considered as an And gate. This will model a pair of required Key Items. The intent is to then update the analyzer to do that, and to also consider multiple [GatedBy] attributes as an Or condition, to model Magma and Hook gating.

This commit currently leaves in and active the failing test that made it clear what was wrong, as well as additional tests that are failing because of the incomplete work.